- object UUID uint32_t -> uint64_t ye - scene load aşamasında pass 2 tamamlandı. Artık parent-child relationship i save/load edebiliyoruz

+++ ModelViewer/Object.cpp
@@ -29,6 +29,6 @@ void Object::serialize(YAML::Emitter& out)
 
 void Object::deserialize(const YAML::Node& node)
 {
-    UUID = node["UUID"].as<uint32_t>();
+    UUID = node["UUID"].as<uint64_t>();
 	name = node["name"].as<std::string>();
 }
